Split Timezone module tests for tasks to new file
Related: rhbz#2127473

(cherry picked from commit 26f51924f5aa48ed907d0cbbed98b5cd20854481)
